move /skill-card into the /shared directory
