+++ src/app/my-skills/my-skills.component.ts
@@ -5,7 +5,7 @@ import { filter, map } from 'rxjs/operators';
 import { MySkillsService } from './my-skills.service';
 import { MySkillsNewComponent } from './my-skills-new.component';
 import { MySkillsEditComponent} from './my-skills-edit.component';
-import { UserSkillView as EditUserSkillView} from '../skill-card/user-skill-view';
+import { UserSkillView as EditUserSkillView} from '../shared/skill-card/user-skill-view';
 import { HttpErrorResponse } from '@angular/common/http';
 import { GlobalErrorHandlerService } from '../error/global-error-handler.service';
 import { DeleteConfirmationDialogComponent } from '../shared/delete-confirmation-dialog/delete-confirmation-dialog.component';
