+++ src/app/shared/skill-card/skill-card.component.html
