+++ src/app/shared/skill-card/skill-card.component.ts
