+++ src/app/shared/skill-card/skill-view.ts
