+++ src/app/shared/skill-card/user-skill-view.ts
